Search Screen Selection and Segues
Fixed an error in the search screen that caused the first item in the
list to always be selected.
Set segues from screens with new board selection to “modal”, this
prevents backward navigation after a new board has been selected.
Made some changes to the organization of code, namely, made the all
board query a function in the Board class.
Added a Board class function to check if the device has an internet
connection, this function is called before a query/scan/update is
initiated, added a new query status code to represent no internet
connectivity.