Email Notifications
Email notifications will be sent when a moderator approves or denies a
post.
Created an email account “bluetoothbillboard@gmail.com” to use as an
application account.
Fixed an issue in the phone number class.
The phone number class is once again used to parse phone numbers form
the post screen.
Grouped files into relevant folders to make the workspace a bit more
manageable.