Background Beacon Scanning
Beacons can now be scanned for in the foreground or background.
Beacon regions are now the object of scanning instead of peripherals.
When beacon regions are found, they are stored in a user default array
that is iterated when viewing the board list screen.
Fixed a minor conversion issue in the phone number class that was using
an improper type.