Notifications and Indicators
Whenever a new beacon is found, a local notification is sent to the
device and the application icon badge is incremented.  Whenever the
application is opened the badge is reset to zero.  Added activity view
indicators to the views with tables, and added constraints to the table
views so they should now size appropriately.  Added a new splash screen
that properly displays the application icon and product owners.