Ban List, Alert Boxes, Validation, and Navigation
Added a new screen named ‘Ban List’.  From the Ban List screen users
will be able to view hosts or types that they have previously banned,
and un-ban them if they desire.
Added numerous alert boxes wherever it was appropriate.  Most of these
boxes simply inform the user that their process was completed or not.
Added some humanity verification in the form of switches on the post
screens and moderator login screen.  These switches are paired with
labels that say “I am not a robot”, in an effort to prevent bots from
abusing the application.
The Moderator login screen now keeps track of failed attempts, and will
disable the login button after 5 fails (not sure when to reactivate it
yet).  The user is notified via alert box for every bad attempt, and a
counter is displayed notably to the user as well.
Several buttons have new segues attached to them now.  Almost none of
the remaining buttons allow the user to remain on the screen they are
currently on (unless it made sense to do that), to avoid redundant
clicking.
Constraints were added to every remaining screen without them.  Views
should now display properly regardless of the device they are being
viewed with.