Redesign Post Viewing
Streamlined all the screens that allow for easier (imo) viewing of
posts (view post, preview post, view queue, and view saved).
Added some validation to the post screen, the user is now required to
enter a host name, post type, and some information about the post, and
will be alerted if not done and not allowed to proceed.  If the user
does not enter a date, a date 90 days (easily changeable) from the date
of submittal will be automatically generated.