Gesture Recognizers
Added long tap gesture recognizers to the saved boards and posts list
screens.  Whenever a user commits a long tap gesture on any cell in one
of those table, they are prompted with a choice to delete that item
from their device.  This is the only method of deleting a board, post
can be deleted either in this fashion or while viewing a saved post.
Fixed some minor constraint issues that were causing buttons to
improperly display.