Outdated Post Removal and UI Tuning
Every time a moderator logs into a board, every post on that board will
be checked to see if its end date is before today, and if it is the
post is deleted.
Fixed an issue with almost every screen where an extra navigation bar
was added erroneously.  All screens seem to be displaying properly now.