More Gestures
Added long tap gestures to several screens.  Long tapping on the search
or board list screens now brings up an alert window with board
information and an option to save the board.  Similar functionality was
added to the view post list screen.
Fixed an error in several locations where null strings were not
evaluating properly.
Changed the way posts are identified slightly.  Now when a post is
submitted, the board is queried to find if the id is already in use.
In order to prevent multiple posts with the same id from different
boards, saving a post now appends the board number onto the post id.
Theoretically it should now be impossible to have a post id collision.
Fixed an issue that executed the submit post segue improperly.
All view controllers are now embedded in navigation controllers, this
should correct the issue of navbars not displaying properly (again).