HTML Viewing
HTML is now usable and viewable in posts.  In an effort to preserve the
formatting of posts without HTML, text views are still used unless
there are HTML tags found in the post information.  When said tags are
found a web view is created to display the message.  To prevent
multiple nested window views, any links selected in the web views will
open the devices browser.